mv table_data to diff file
